+++ infoschema/cluster.go
@@ -104,6 +104,7 @@ func AppendHostInfoToRows(ctx sessionctx.Context, rows [][]types.Datum) ([][]typ
 	return rows, nil
 }
 
+// GetInstanceAddr gets the instance address.
 func GetInstanceAddr(ctx sessionctx.Context) (string, error) {
 	serverInfo, err := infosync.GetServerInfo()
 	if err != nil {
