+++ util/stmtsummary/statement_summary_test.go
@@ -624,7 +624,7 @@ func generateAnyExecInfo() *StmtExecInfo {
 	return stmtExecInfo
 }
 
-func newStmtSummaryReaderForTest(ssMap *stmtSummaryByDigestMap) *StmtSummaryReader {
+func newStmtSummaryReaderForTest(ssMap *stmtSummaryByDigestMap) *stmtSummaryReader {
 	columnNames := []string{
 		SummaryBeginTimeStr,
 		SummaryEndTimeStr,
@@ -702,7 +702,7 @@ func newStmtSummaryReaderForTest(ssMap *stmtSummaryByDigestMap) *StmtSummaryRead
 		AvgKvTimeStr,
 		AvgPdTimeStr,
 		AvgBackoffTotalTimeStr,
-		AvgWriteSqlRespTimeStr,
+		AvgWriteSQLRespTimeStr,
 		PreparedStr,
 		AvgAffectedRowsStr,
 		FirstSeenStr,
@@ -738,7 +738,7 @@ func (s *testStmtSummarySuite) TestToDatum(c *C) {
 	stmtExecInfo1 := generateAnyExecInfo()
 	s.ssMap.AddStatement(stmtExecInfo1)
 	reader := newStmtSummaryReaderForTest(s.ssMap)
-	datums := reader.GetAllStmtSummaryRows()
+	datums := reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, 1)
 	n := types.NewTime(types.FromGoTime(time.Unix(s.ssMap.beginTimeForCurInterval, 0)), mysql.TypeTimestamp, types.DefaultFsp)
 	e := types.NewTime(types.FromGoTime(time.Unix(s.ssMap.beginTimeForCurInterval+1800, 0)), mysql.TypeTimestamp, types.DefaultFsp)
@@ -774,7 +774,7 @@ func (s *testStmtSummarySuite) TestToDatum(c *C) {
 		t, t, 0, 0, 0, stmtExecInfo1.OriginalSQL, stmtExecInfo1.PrevSQL, "plan_digest", ""}
 	stmtExecInfo1.ExecDetail.CommitDetail.Mu.Unlock()
 	match(c, datums[0], expectedDatum...)
-	datums = reader.GetAllStmtSummaryHistoryRows()
+	datums = reader.GetStmtSummaryHistoryRows()
 	c.Assert(len(datums), Equals, 1)
 	match(c, datums[0], expectedDatum...)
 
@@ -791,7 +791,7 @@ func (s *testStmtSummarySuite) TestToDatum(c *C) {
 	stmtExecInfo2.Digest = "bandit sei"
 	s.ssMap.AddStatement(stmtExecInfo2)
 	c.Assert(s.ssMap.summaryMap.Size(), Equals, 1)
-	datums = reader.GetAllStmtSummaryRows()
+	datums = reader.GetStmtSummaryCurrentRows()
 	expectedEvictedDatum := []interface{}{n, e, "", "<nil>", "<nil>", "",
 		"<nil>", "<nil>", stmtExecInfo1.User, 1, 0, 0, int64(stmtExecInfo1.TotalLatency),
 		int64(stmtExecInfo1.TotalLatency), int64(stmtExecInfo1.TotalLatency), int64(stmtExecInfo1.TotalLatency),
@@ -849,7 +849,7 @@ func (s *testStmtSummarySuite) TestAddStatementParallel(c *C) {
 		}
 
 		// There would be 32 summaries.
-		datums := reader.GetAllStmtSummaryRows()
+		datums := reader.GetStmtSummaryCurrentRows()
 		c.Assert(len(datums), Equals, loops)
 	}
 
@@ -858,7 +858,7 @@ func (s *testStmtSummarySuite) TestAddStatementParallel(c *C) {
 	}
 	wg.Wait()
 
-	datums := reader.GetAllStmtSummaryRows()
+	datums := reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, loops)
 }
 
@@ -1002,7 +1002,7 @@ func (s *testStmtSummarySuite) TestSetMaxStmtCountParallel(c *C) {
 	wg.Wait()
 
 	reader := newStmtSummaryReaderForTest(s.ssMap)
-	datums := reader.GetAllStmtSummaryRows()
+	datums := reader.GetStmtSummaryCurrentRows()
 	// due to evictions happened in cache, an additional record will be appended to the table.
 	c.Assert(len(datums), Equals, 2)
 }
@@ -1020,7 +1020,7 @@ func (s *testStmtSummarySuite) TestDisableStmtSummary(c *C) {
 	stmtExecInfo1 := generateAnyExecInfo()
 	s.ssMap.AddStatement(stmtExecInfo1)
 	reader := newStmtSummaryReaderForTest(s.ssMap)
-	datums := reader.GetAllStmtSummaryRows()
+	datums := reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, 0)
 
 	// Set true in session scope, it will overwrite global scope.
@@ -1028,7 +1028,7 @@ func (s *testStmtSummarySuite) TestDisableStmtSummary(c *C) {
 	c.Assert(err, IsNil)
 
 	s.ssMap.AddStatement(stmtExecInfo1)
-	datums = reader.GetAllStmtSummaryRows()
+	datums = reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, 1)
 
 	// Set false in global scope, it shouldn't work.
@@ -1041,7 +1041,7 @@ func (s *testStmtSummarySuite) TestDisableStmtSummary(c *C) {
 	stmtExecInfo2.NormalizedSQL = "normalized_sql2"
 	stmtExecInfo2.Digest = "digest2"
 	s.ssMap.AddStatement(stmtExecInfo2)
-	datums = reader.GetAllStmtSummaryRows()
+	datums = reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, 2)
 
 	// Unset in session scope.
@@ -1049,7 +1049,7 @@ func (s *testStmtSummarySuite) TestDisableStmtSummary(c *C) {
 	c.Assert(err, IsNil)
 	s.ssMap.beginTimeForCurInterval = now + 60
 	s.ssMap.AddStatement(stmtExecInfo2)
-	datums = reader.GetAllStmtSummaryRows()
+	datums = reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, 0)
 
 	// Unset in global scope.
@@ -1057,7 +1057,7 @@ func (s *testStmtSummarySuite) TestDisableStmtSummary(c *C) {
 	c.Assert(err, IsNil)
 	s.ssMap.beginTimeForCurInterval = now + 60
 	s.ssMap.AddStatement(stmtExecInfo1)
-	datums = reader.GetAllStmtSummaryRows()
+	datums = reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, 1)
 
 	// Set back.
@@ -1086,7 +1086,7 @@ func (s *testStmtSummarySuite) TestEnableSummaryParallel(c *C) {
 			c.Assert(err, IsNil)
 			s.ssMap.AddStatement(stmtExecInfo1)
 			// Try to read it.
-			reader.GetAllStmtSummaryHistoryRows()
+			reader.GetStmtSummaryHistoryRows()
 		}
 		err := s.ssMap.SetEnabled("1", false)
 		c.Assert(err, IsNil)
@@ -1222,12 +1222,12 @@ func (s *testStmtSummarySuite) TestSummaryHistory(c *C) {
 		}
 	}
 	reader := newStmtSummaryReaderForTest(s.ssMap)
-	datum := reader.GetAllStmtSummaryHistoryRows()
+	datum := reader.GetStmtSummaryHistoryRows()
 	c.Assert(len(datum), Equals, 10)
 
 	err = s.ssMap.SetHistorySize("5", false)
 	c.Assert(err, IsNil)
-	datum = reader.GetAllStmtSummaryHistoryRows()
+	datum = reader.GetStmtSummaryHistoryRows()
 	c.Assert(len(datum), Equals, 5)
 
 	// test eviction
@@ -1252,7 +1252,7 @@ func (s *testStmtSummarySuite) TestSummaryHistory(c *C) {
 	c.Assert(s.ssMap.summaryMap.Size(), Equals, 1)
 	// length of `other` should not longer than historySize.
 	c.Assert(s.ssMap.other.history.Len(), Equals, 5)
-	datum = reader.GetAllStmtSummaryHistoryRows()
+	datum = reader.GetStmtSummaryHistoryRows()
 	// length of STATEMENT_SUMMARY_HISTORY == (history in cache) + (history evicted)
 	c.Assert(len(datum), Equals, 6)
 }
@@ -1380,27 +1380,27 @@ func (s *testStmtSummarySuite) TestAccessPrivilege(c *C) {
 	badUser := &auth.UserIdentity{Username: "bad_user"}
 
 	reader := newStmtSummaryReaderForTest(s.ssMap)
-	reader.User = user
+	reader.user = user
 	reader.isSuper = false
-	datums := reader.GetAllStmtSummaryRows()
+	datums := reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, loops)
-	reader.User = badUser
+	reader.user = badUser
 	reader.isSuper = false
-	datums = reader.GetAllStmtSummaryRows()
+	datums = reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, 0)
 	reader.isSuper = true
-	datums = reader.GetAllStmtSummaryRows()
+	datums = reader.GetStmtSummaryCurrentRows()
 	c.Assert(len(datums), Equals, loops)
 
-	reader.User = user
+	reader.user = user
 	reader.isSuper = false
-	datums = reader.GetAllStmtSummaryHistoryRows()
+	datums = reader.GetStmtSummaryHistoryRows()
 	c.Assert(len(datums), Equals, loops)
-	reader.User = badUser
+	reader.user = badUser
 	reader.isSuper = false
-	datums = reader.GetAllStmtSummaryHistoryRows()
+	datums = reader.GetStmtSummaryHistoryRows()
 	c.Assert(len(datums), Equals, 0)
 	reader.isSuper = true
-	datums = reader.GetAllStmtSummaryHistoryRows()
+	datums = reader.GetStmtSummaryHistoryRows()
 	c.Assert(len(datums), Equals, loops)
 }
